remove stopSound and some changes
