+++ src/chrome/handleWork.js
@@ -1,10 +1,9 @@
 import listenStoreChanges from '../listenStoreChanges';
 import { workFinished } from '../actions';
-import { playSoundOptions, stopSound } from '../playSound';
+import { playSound } from '../playSound';
 
 const handleWork = (store, chrome) => {
   let workTabId = null;
-  let audioPlayer = null;
 
   const handleAssignWork = ({ worker: prevWorker }, { worker: curWorker }) => {
     if (prevWorker.get('state') !== 'working' && curWorker.get('state') === 'working') {
@@ -16,15 +15,11 @@ const handleWork = (store, chrome) => {
       const oldWorkTabId = workTabId;
       chrome.tabs.create({ url }, tab => {
         workTabId = tab.id;
-        audioPlayer = playSoundOptions(store.getState().plugin.get('options'));
+        playSound(store.getState().plugin.get('options'));
       });
       if (oldWorkTabId) {
         chrome.tabs.remove(oldWorkTabId);
       }
-    }
-
-    if (prevWorker.get('state') === 'working' && curWorker.get('state') !== 'working') {
-      stopSound(audioPlayer);
     }
   };
